more data methods to indexing
